requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.0.2
boto3==1.9.94
botocore==1.12.94
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cmake==3.13.2.post1
colorama==0.4.0
cycler==0.10.0
Cython==0.29.1
dask==1.0.0
decorator==4.3.0
demjson==2.2.4
device==0.1
Django==2.1.2
django-mysql==2.4.1
dlib==19.16.0
docutils==0.14
docx==0.2.4
entrypoints==0.2.3
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
future==0.17.1
gast==0.2.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-dataproc==0.3.0
google-cloud-logging==1.10.0
google-cloud-speech==0.36.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
grpcio==1.12.1
h5py==2.8.0
idna==2.8
imageio==2.4.1
implicit==0.3.7
imutils==0.5.2
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
lxml==4.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mrjob==0.6.7
mysql-connector-python==8.0.13
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.3
notebook==5.7.2
numpy==1.15.4
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
phe==1.4.0
pickleshare==0.7.5
Pillow==5.3.0
pocketsphinx==0.1.15
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pycparser==2.19
Pygments==2.2.0
pyparsing==2.3.0
pyserial==3.4
pytesseract==0.2.5
python-dateutil==2.7.5
python-docx==0.8.10
python-speech-features==0.6
pytz==2018.7
PyVCF==0.6.8
PyWavelets==1.0.1
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
R305==2.0.0
requests==2.0.0
rsa==4.0
s3transfer==0.2.0
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
Send2Trash==1.5.0
serial==0.0.97
six==1.11.0
sklearn==0.0
sounddevice==0.3.12
soupsieve==1.7.3
SpeechRecognition==3.8.1
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
tesseract==0.1.3
tesserocr==2.2.2
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24.1
wavio==0.0.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wincertstore==0.2
